_admin/requirements.txt
alembic==0.8.3
amqp==1.4.7
anyjson==0.3.3
Babel==2.1.1
billiard==3.3.0.20
blinker==1.4
celery==3.1.18
defusedxml==0.4.1
Flask==0.10.1
Flask-Babel==0.9
Flask-Cache==0.13.1
Flask-FlatPages==0.6
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==1.6.0
Flask-Restless==0.17.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.28
Mako==1.0.2
Markdown==2.6.2
MarkupSafe==0.23
mimerender==0.5.5
oauthlib==1.0.3
PyJWT==1.4.0
python-dateutil==2.4.2
python-editor==0.4
python-mimeparse==0.1.4
python-social-auth==0.2.1
python3-openid==3.0.7
pytz==2015.6
PyYAML==3.11
redis==2.10.3
requests==2.8.1
requests-oauthlib==0.3.1
six==1.10.0
speaklater==1.3
SQLAlchemy==1.0.9
Werkzeug==0.10.4
wheel==0.24.0
WTForms==2.0.2
setuptools==18.4
psycopg2
trafaret
PyGithub==1.25.2
simplejson
mailchimp==2.0.9
